Sketch.js
Added a Boost percentage. 

You can use the boost once it reaches 100.

Press ( X ) or ( , ) to activate it depending on your respective side.
Players will only have 1 chance when it comes to using it, after the next ball bounce, you will lose it.

To gain percentage, you have to bounce it on your paddle 5 times for 5 percent, or score for 12 percent.